api/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
snowflake-connector-python
python-dotenv
